range-diff: fix integer overflow in get_correspondences()
The get_correspondences() function uses 'int' to store the sum of two
'size_t' values from string_list structures. When processing large
patch sets where a->nr + b->nr exceeds INT_MAX, the value wraps to
negative, causing invalid array indexing and a segmentation fault.

This manifests as a crash at line 356 in range-diff.c when accessing
cost[i + n * j] with a negative or incorrectly calculated index.

Fix this by using 'size_t' throughout for array sizes and indices.
The compute_assignment() function signature is updated to accept
size_t parameters, maintaining int only for sentinel values (-1).

Note that while this fix prevents the integer overflow and segmentation
fault, attempting to process extremely large patch sets will still hit
memory limitations. In practice, the process will consume excessive
memory and likely be terminated by the system (SIGKILL) before
completing. However, fixing the integer overflow is still correct and
allows range-diff to fail gracefully due to resource constraints rather
than undefined behavior.

+++ range-diff.c
@@ -327,9 +327,9 @@ static int diffsize(const char *a, const char *b)
 static void get_correspondences(struct string_list *a, struct string_list *b,
 				int creation_factor)
 {
-	int n = a->nr + b->nr;
+	size_t n = a->nr + b->nr;
 	int *cost, c, *a2b, *b2a;
-	int i, j;
+	size_t i, j;
 
 	ALLOC_ARRAY(cost, st_mult(n, n));
 	ALLOC_ARRAY(a2b, n);
